Move SecureBackupConfig to the appconfig module.
