+++ features/securebackup/impl/src/main/kotlin/io/element/android/features/securebackup/impl/root/SecureBackupRootNode.kt
@@ -27,7 +27,7 @@ import com.bumble.appyx.core.plugin.plugins
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
 import io.element.android.anvilannotations.ContributesNode
-import io.element.android.features.securebackup.impl.SecureBackupConfig
+import io.element.android.appconfig.SecureBackupConfig
 import io.element.android.libraries.di.SessionScope
 
 @ContributesNode(SessionScope::class)
